+++ src/Function.ts
@@ -839,3 +839,22 @@ export const applySomes =
       fs,
       A.reduce(x, (y, mf) => pipe(mf, O.match(constant(y), apply(y)))),
     )
+
+/**
+ * Apply a function the specified number of times. `n` will be normalised to a
+ * non-negative integer according to the semantics of `A.replicate`.
+ *
+ * @example
+ * import { applyN } from 'fp-ts-std/Function'
+ * import { increment } from 'fp-ts-std/Number'
+ *
+ * const add3 = applyN(3)(increment)
+ *
+ * assert.strictEqual(add3(2), 5)
+ *
+ * @since 0.17.0
+ */
+export const applyN =
+  (n: number) =>
+  <A>(f: Endomorphism<A>): Endomorphism<A> =>
+    pipe(A.replicate(n, f), applyEvery)